LayoutTests/platform/ios-simulator/editing/deleting/delete-surrogatepair-expected.txt

This test makes sure that a surrogate-pair character is deleted by one pressing delete key

hogehoge
𠀾hogehoge
a𠀾
a𠀾
𠀾a
𠀾a
a𠀾a
aa
a𠀾
a𠀾𠀾a
a𠀾𠀾a
